categorize windows.unhooked_system_calls as a malware plugin
